Shell+Editor+Zsh: Switch to fzf-marks from vimwiki
So far I have kept a `wiki.vim` file (which for the most part I have not
committed in order to not share the details of my FS hierarchy). As it
has a major drawback (actually a couple including the longevity of the
"configuration file", the private nature of the files which I want to
"index", the markdown vs. wiki battle, etc.) I have been looking for an alternative for quite some
time now.

Today I stumbled upon this great tool called fzf-marks, which as the
name suggests, uses one of my favourite tools - fzf. It is incredibly
simple - as all such utilities should be. It keeps a file referenced as
`FZF_MARKS_FILE` and uses it to jump around to these "shortcuts". This
thing automatically removes by need of `vimwiki` for me as my usual
approach is 1) Go to this special "wiki" directory, 2) use FZF to find
the files I need. Moreover, it removes the "changing directory" section
of my `shortcuts.sh` file.

GJ! I am pretty sure I am going to enjoy the use of it :).

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>